Changes to be committed: 	modified:   boards/licheerv_nano/stupid_fly/src/board_config.h 	modified:   platforms/common/CMakeLists.txt 	modified:   platforms/nuttx/NuttX/nuttx 	modified:   src/drivers/uavcan/CMakeLists.txt 	modified:   src/drivers/uavcan/sensors/fuel_tank_status.cpp 	modified:   src/drivers/uavcan/uavcan_drivers/sg20xx/driver/include/uavcan_sg20xx/can.hpp 	modified:   src/drivers/uavcan/uavcan_drivers/sg20xx/driver/src/uc_sg2002_can.cpp 	modified:   src/drivers/uavcan/uavcan_main.cpp 	modified:   src/modules/commander/HealthAndArmingChecks/checks/airspeedCheck.cpp 	modified:   src/modules/commander/HealthAndArmingChecks/checks/estimatorCheck.cpp 	modified:   src/modules/commander/HealthAndArmingChecks/checks/systemCheck.hpp 	modified:   src/modules/control_allocator/ControlAllocator.cpp 	modified:   src/modules/mavlink/streams/VFR_HUD.hpp
1. update

+++ src/drivers/uavcan/sensors/fuel_tank_status.cpp
@@ -80,7 +80,7 @@ void UavcanFuelTankStatusBridge::fuel_tank_status_sub_cb(const
 	report.fuel_tank_id = msg.fuel_tank_id;
 
 	// Optional temperature field, in Kelvin, set to NaN if not provided.
-	report.temperature = !PX4_ISFINITE(msg.fuel_temperature) ? NAN : msg.fuel_temperature;
+	report.temperature = !PX4_ISFINITE(msg.fuel_temperature) ? NAN : static_cast<double>(msg.fuel_temperature);
 
 	publish(msg.getSrcNodeID().get(), &report);
 }
