+++ src/modules/mavlink/streams/VFR_HUD.hpp
@@ -88,7 +88,7 @@ private:
 
 			mavlink_vfr_hud_t msg{};
 			// display NAN in case of source not being one of the sensors
-			msg.airspeed = airspeed_from_sensor ? airspeed_validated.calibrated_airspeed_m_s : NAN;
+			msg.airspeed = airspeed_from_sensor ? airspeed_validated.calibrated_airspeed_m_s : static_cast<float>(NAN);
 			msg.groundspeed = sqrtf(lpos.vx * lpos.vx + lpos.vy * lpos.vy);
 			msg.heading = math::degrees(matrix::wrap_2pi(lpos.heading));
